Refactored Staging add_account into a struct method
